Check for localStorage support so IE7 doesn't choke

+++ js/saiku/models/Session.js
@@ -14,12 +14,12 @@ var Session = Backbone.Model.extend({
         _.bindAll(this, "process_login", "prefetch_dimensions");
         
         // Check expiration on localStorage
-        if (! (localStorage.getItem('expiration') > (new Date()).getTime())) {
+        if (localStorage && ! (localStorage.getItem('expiration') > (new Date()).getTime())) {
             localStorage.clear();
         }
         
         // Check if credentials are already stored
-        if (sessionStorage) {
+        if (localStorage) {
             this.username = localStorage.getItem('username');
             this.password = localStorage.getItem('password');
         }
@@ -55,14 +55,14 @@ var Session = Backbone.Model.extend({
     
     login: function(username, password) {
         this.username = username;
-        localStorage.setItem('username', username);
+        localStorage && localStorage.setItem('username', username);
         this.password = password;
-        localStorage.setItem('password', password);
+        localStorage && localStorage.setItem('password', password);
         
         // Set expiration on localStorage to one day in the future
         var expires = (new Date()).getTime() + 
             Settings.LOCALSTORAGE_EXPIRATION;
-        localStorage.setItem('expiration', expires);
+        localStorage && localStorage.setItem('expiration', expires);
         
         // Create session and fetch connection information
         this.fetch({ success: this.process_login });
@@ -76,7 +76,7 @@ var Session = Backbone.Model.extend({
     logout: function() {
         // FIXME - This is a hack (inherited from old UI)
         $('body').hide();
-        localStorage.clear();
+        localStorage && localStorage.clear();
         location.reload(true);
         return false;
     },
